[Savestates] Fixed client motion not working
